Optimize Heater to eliminate missed Serial characters
Disabling interrupts then retrieving the latest temperature was (somehow) taking too long. Perhaps the memory allocations?
It's not necessary to disable interrupts to retrieve that value, so removing the 'disable' fixed the missing characters (bad checksums) issue.

Also tweaked the bounds check to return if out-of-bound...otherwise the heater control code could turn the heater back on.