Stepper - eliminate possible null dereference when quickStop() called
Now that we allow other interrupts during the stepper isr, it's possible that quickStop() is called, sets current_block to null, then the stepper attempts to dereference that var resulting in undefined behaviour. Instead, we use a flag, and allow the isr to clear the current and any pending blocks.